Fix four API mismatches found by compiling against a real headset
Building against the IL2CPP assemblies pulled from a Quest 3S copy of BONELAB
surfaced four differences from what Unity's documentation implies. None of them
would have failed a build against the hand-written references, which is exactly
why that pass was worth doing.

- FindObjectsOfType<T>() returns Il2CppArrayBase<T>, not
  Il2CppReferenceArray<T>. Declare the type it actually returns rather than
  downcasting to a subtype we were only assuming.
- GarbageCollector.incrementalTimeSliceNanoseconds is ulong, not uint.
- ILogger.logEnabled is projected read-only, so the game's logger cannot be
  switched off this way. Rather than guess at another route, drop the
  DisableUnityLogger option. It was off by default, marginal next to stack trace
  stripping, and its own description warned that it hid real errors - so the
  honest fix is to stop offering it.

The reference assemblies now carry the verified signatures instead of the
guesses, so CI validates against what the game really exposes.

+++ src/Tweaks/MemoryTweak.cs
@@ -25,7 +25,7 @@ namespace Parity.Tweaks
         public override string Name => "Memory";
 
         private bool _captured;
-        private uint _originalSliceNanoseconds;
+        private ulong _originalSliceNanoseconds;
         private bool _warnedNotIncremental;
 
         public override void Apply()
@@ -55,7 +55,7 @@ namespace Parity.Tweaks
                     _captured = true;
                 }
 
-                uint want = _originalSliceNanoseconds;
+                ulong want = _originalSliceNanoseconds;
 
                 if (ParityPreferences.TuneIncrementalGc.Value)
                 {
@@ -63,7 +63,7 @@ namespace Parity.Tweaks
                         ParityPreferences.IncrementalGcSliceMicroseconds.Value,
                         MinSliceMicroseconds,
                         MaxSliceMicroseconds);
-                    want = (uint)microseconds * 1000u;
+                    want = (ulong)microseconds * 1000ul;
                 }
 
                 if (GarbageCollector.incrementalTimeSliceNanoseconds != want)
